+++ src/Render/types/index.ts
@@ -1,10 +1,5 @@
 // 配置相关
-export type {
-	ZoomConfig,
-	RenderConfig,
-	DropMaterialData,
-	OnDropCallback,
-} from "./config";
+export type { ZoomConfig, RenderConfig } from "./config";
 
 // 实例相关
 export type {