+++ src/Render/utils/index.ts
@@ -1,0 +1,8 @@
+export {
+	loadSvg,
+	loadGif,
+	loadImg,
+	loadAsset,
+	preloadAssets,
+	playGif,
+} from "./assetLoader";